requirements.txt
alembic==1.5.8
click==7.1.2
cryptography==35.0.0
decorator==4.4.2
dnspython==2.1.0
email-validator==1.1.1
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Login==0.5.0
flask-mailman==0.1.7
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-Reuploaded==0.3.2
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==3.1
infinity==1.5
intervals==0.9.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.23.1
Pillow==8.0.1
PyMySQL==1.0.2
python-dateutil==2.8.1
python-editor==1.0.4
shopyo==3.9.4
six==1.15.0
SQLAlchemy==1.3.23
SQLAlchemy-Utils==0.36.8
validators==0.18.2
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.5
tzdata
backports.zoneinfo